+++ caffe2/core/tensor_impl.cc
@@ -1,0 +1,14 @@
+#include "caffe2/core/tensor_impl.h"
+
+#include "caffe2/core/flags.h"
+
+CAFFE2_DEFINE_bool(
+    caffe2_keep_on_shrink,
+    true,
+    "If set, keeps memory when a tensor is shrinking its size.");
+
+CAFFE2_DEFINE_int64(
+    caffe2_max_keep_on_shrink_memory,
+    LLONG_MAX,
+    "The maximum memory in bytes to keep on shrink, if the difference between "
+    "tensor sizes is bigger than this then tensor will be reset.");
